Reorganise the content
