+++ pkg/registry/provider_embedded.go
