+++ pkg/registry/provider_remote.go
