--- pkg/registry/registry.go
@@ -1,2 +1,0 @@
-package registry
-
