--- pkg/registry/registry_test.go
@@ -1,119 +1,0 @@
-package registry
-
-import (
-	"testing"
-)
-
-func TestGetRegistry(t *testing.T) {
-	t.Parallel()
-	provider, err := GetDefaultProvider()
-	if err != nil {
-		t.Fatalf("Failed to get registry provider: %v", err)
-	}
-	reg, err := provider.GetRegistry()
-	if err != nil {
-		t.Fatalf("Failed to get registry: %v", err)
-	}
-
-	if reg == nil {
-		t.Fatal("Registry is nil")
-		return
-	}
-
-	if reg.Version == "" {
-		t.Error("Registry version is empty")
-	}
-
-	if reg.LastUpdated == "" {
-		t.Error("Registry last updated is empty")
-	}
-
-	if len(reg.Servers) == 0 {
-		t.Error("Registry has no servers")
-	}
-}
-
-func TestGetServer(t *testing.T) {
-	t.Parallel()
-	// Test getting an existing server
-	provider, err := GetDefaultProvider()
-	if err != nil {
-		t.Fatalf("Failed to get registry provider: %v", err)
-	}
-	server, err := provider.GetServer("brave-search")
-	if err != nil {
-		t.Fatalf("Failed to get server: %v", err)
-	}
-
-	if server == nil {
-		t.Fatal("ImageMetadata is nil")
-		return
-	}
-
-	if server.Image == "" {
-		t.Error("ImageMetadata image is empty")
-	}
-
-	if server.Description == "" {
-		t.Error("ImageMetadata description is empty")
-	}
-
-	// Test getting a non-existent server
-	_, err = provider.GetServer("non-existent-server")
-	if err == nil {
-		t.Error("Expected error when getting non-existent server")
-	}
-}
-
-func TestSearchServers(t *testing.T) {
-	t.Parallel()
-	// Test searching for servers
-	provider, err := GetDefaultProvider()
-	if err != nil {
-		t.Fatalf("Failed to get registry provider: %v", err)
-	}
-	servers, err := provider.SearchServers("search")
-	if err != nil {
-		t.Fatalf("Failed to search servers: %v", err)
-	}
-
-	if len(servers) == 0 {
-		t.Error("No servers found for search query")
-	}
-
-	// Test searching for non-existent servers
-	servers, err = provider.SearchServers("non-existent-server")
-	if err != nil {
-		t.Fatalf("Failed to search servers: %v", err)
-	}
-
-	if len(servers) > 0 {
-		t.Errorf("Expected no servers for non-existent query, got %d", len(servers))
-	}
-}
-
-func TestListServers(t *testing.T) {
-	t.Parallel()
-	provider, err := GetDefaultProvider()
-	if err != nil {
-		t.Fatalf("Failed to get registry provider: %v", err)
-	}
-	servers, err := provider.ListServers()
-	if err != nil {
-		t.Fatalf("Failed to list servers: %v", err)
-	}
-
-	if len(servers) == 0 {
-		t.Error("No servers found")
-	}
-
-	// Verify that we get the same number of servers as in the registry
-	reg, err := provider.GetRegistry()
-	if err != nil {
-		t.Fatalf("Failed to get registry: %v", err)
-	}
-
-	if len(servers) != len(reg.Servers) {
-		t.Errorf("Expected %d servers, got %d", len(reg.Servers), len(servers))
-	}
-}
